# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/health": {
            "get": {
                "description": "Check if service is healthy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/order/{order_uid}": {
            "get": {
                "description": "Get order details by order_uid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by UID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order UID",
                        "name": "order_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Delivery": {
            "type": "object",
            "required": [
                "address",
                "city",
                "email",
                "name",
                "phone",
                "region",
                "zip"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                },
                "city": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "phone": {
                    "type": "string"
                },
                "region": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "zip": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "domain.Item": {
            "type": "object",
            "required": [
                "brand",
                "chrt_id",
                "name",
                "nm_id",
                "price",
                "rid",
                "size",
                "status",
                "track_number"
            ],
            "properties": {
                "brand": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "chrt_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "nm_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "price": {
                    "type": "integer",
                    "minimum": 1
                },
                "rid": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 10
                },
                "sale": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "size": {
                    "type": "string",
                    "maxLength": 50
                },
                "status": {
                    "type": "integer",
                    "maximum": 600,
                    "minimum": 100
                },
                "total_price": {
                    "type": "integer",
                    "minimum": 0
                },
                "track_number": {
                    "type": "string"
                }
            }
        },
        "domain.Order": {
            "type": "object",
            "required": [
                "customer_id",
                "date_created",
                "delivery",
                "delivery_service",
                "entry",
                "items",
                "locale",
                "oof_shard",
                "order_uid",
                "payment",
                "shardkey",
                "sm_id",
                "track_number"
            ],
            "properties": {
                "customer_id": {
                    "type": "string",
                    "maxLength": 50
                },
                "date_created": {
                    "type": "string"
                },
                "delivery": {
                    "$ref": "#/definitions/domain.Delivery"
                },
                "delivery_service": {
                    "type": "string",
                    "maxLength": 50
                },
                "entry": {
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 3
                },
                "internal_signature": {
                    "type": "string",
                    "maxLength": 255
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.Item"
                    }
                },
                "locale": {
                    "type": "string",
                    "enum": [
                        "en",
                        "ru"
                    ]
                },
                "oof_shard": {
                    "type": "string",
                    "maxLength": 10
                },
                "order_uid": {
                    "type": "string"
                },
                "payment": {
                    "$ref": "#/definitions/domain.Payment"
                },
                "shardkey": {
                    "type": "string",
                    "maxLength": 10
                },
                "sm_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "track_number": {
                    "type": "string"
                }
            }
        },
        "domain.Payment": {
            "type": "object",
            "required": [
                "amount",
                "bank",
                "currency",
                "payment_dt",
                "provider",
                "transaction"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "bank": {
                    "type": "string",
                    "maxLength": 100
                },
                "currency": {
                    "type": "string"
                },
                "custom_fee": {
                    "type": "integer",
                    "minimum": 0
                },
                "delivery_cost": {
                    "type": "integer",
                    "minimum": 0
                },
                "goods_total": {
                    "type": "integer",
                    "minimum": 0
                },
                "payment_dt": {
                    "type": "integer",
                    "minimum": 0
                },
                "provider": {
                    "type": "string",
                    "maxLength": 50
                },
                "request_id": {
                    "type": "string",
                    "maxLength": 50
                },
                "transaction": {
                    "type": "string"
                }
            }
        }
    }
}